Optimize implement location
